Fixed tiny error in covariance matrix
The covariance matrix used as an argument to the non-central t CDF
contained 2 tiny errors: the diagonal elements contained a factor of n-3,
and the off-diagonal element contained a factor of n, but both of these
should be n-1. The error in the resulting calculations was totally negligible
for all but the tiniest sample sizes.